requirements.txt
dlib==19.24.2
matplotlib
numpy
opencv_python==4.9.0.80
pandas
Pillow==10.3.0
retina_face==0.0.14
seaborn==0.13.2
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
urllib3==2.2.1
Flask==3.1.0
gunicorn
